+++ ts/components/messages/paginated/MessageList/Item.tsx
@@ -5,7 +5,8 @@ import * as O from "fp-ts/lib/Option";
 import LegalMessage from "../../../../../img/features/mvl/legalMessage.svg";
 import { ServicePublic } from "../../../../../definitions/backend/ServicePublic";
 import { MessageCategory } from "../../../../../definitions/backend/MessageCategory";
-import { TagEnum } from "../../../../../definitions/backend/MessageCategoryBase";
+import { TagEnum as TagEnumBase } from "../../../../../definitions/backend/MessageCategoryBase";
+import { TagEnum as TagEnumPN } from "../../../../../definitions/backend/MessageCategoryPN";
 import I18n from "../../../../i18n";
 import {
   convertDateToWordDistance,
@@ -20,6 +21,7 @@ import IconFont from "../../../ui/IconFont";
 import customVariables from "../../../../theme/variables";
 import { IOColors } from "../../../core/variables/IOColors";
 import QrCode from "../../../../../img/messages/qr-code.svg";
+import { mvlEnabled, pnEnabled } from "../../../../config";
 
 const ICON_WIDTH = 24;
 
@@ -187,10 +189,14 @@ const itemBadgeToAccessibilityLabel = (itemBadge: ItemBadge): string => {
 
 function getTopIcon(category: MessageCategory) {
   switch (category.tag) {
-    case TagEnum.LEGAL_MESSAGE:
-      return (
+    case TagEnumBase.LEGAL_MESSAGE:
+      return mvlEnabled ? (
         <LegalMessage width={20} height={20} fill={IOColors.bluegreyLight} />
-      );
+      ) : null;
+    case TagEnumPN.PN:
+      return pnEnabled ? (
+        <LegalMessage width={20} height={20} fill={IOColors.bluegreyLight} />
+      ) : null;
     default:
       return null;
   }