+++ ts/sagas/startup/loadSessionInformationSaga.ts
@@ -10,6 +10,7 @@ import {
 
 import { BackendClient } from "../../api/backend";
 import { ReduxSagaEffect, SagaCallReturnType } from "../../types/utils";
+import { convertUnknownToError } from "../../utils/errors";
 
 /**
  * Load session info from the Backend
@@ -46,7 +47,7 @@ export function* loadSessionInformationSaga(
       ? response.value.value.title || errorMsgDefault
       : errorMsgDefault;
   } catch (e) {
-    yield* put(sessionInformationLoadFailure(Error(e)));
+    yield* put(sessionInformationLoadFailure(convertUnknownToError(e)));
     return none;
   }
 }